feat(webpack): remove support for legacy browsers (#14190)
