--- packages/react/plugins/storybook/index.spec.ts
@@ -1,33 +1,0 @@
-import { webpack } from './index';
-import { join } from 'path';
-
-jest.mock('@nrwl/webpack/src/executors/webpack/lib/get-webpack-config', () => {
-  return {
-    getStylesPartial: () => ({}),
-  };
-});
-
-describe('Storybook webpack config', () => {
-  it('should skip type checking', async () => {
-    const config = await webpack(
-      {
-        resolve: {
-          plugins: [],
-        },
-        plugins: [],
-        module: {
-          rules: [],
-        },
-      },
-      {
-        configDir: join(__dirname, '../..'),
-      }
-    );
-
-    expect(
-      config.plugins.find(
-        (p) => p.constructor.name === 'ForkTsCheckerWebpackPlugin'
-      )
-    ).toBeFalsy();
-  });
-});
