+++ packages/webpack/index.ts
@@ -1,3 +1,4 @@
+export * from './src/utils/create-copy-plugin';
 export * from './src/utils/config';
 export * from './src/generators/init/init';
 export * from './src/generators/webpack-project/webpack-project';
@@ -11,3 +12,6 @@ export type {
   FileReplacement,
 } from './src/executors/webpack/schema';
 export * from './src/executors/webpack/webpack.impl';
+export * from './src/utils/get-css-module-local-ident';
+export * from './src/utils/with-nx';
+export * from './src/utils/with-web';
