+++ packages/webpack/src/executors/dev-server/dev-server.impl.ts
@@ -30,6 +30,7 @@ export async function* devServerExecutor(
   const buildOptions = normalizeOptions(
     getBuildOptions(serveOptions, context),
     context.root,
+    projectRoot,
     sourceRoot
   );
