+++ packages/webpack/src/plugins/generate-package-json-webpack-plugin.ts
