+++ packages/webpack/src/utils/get-client-environment.ts
@@ -1,0 +1,30 @@
+export function getClientEnvironment(mode?: string) {
+  // Grab NODE_ENV and NX_* environment variables and prepare them to be
+  // injected into the application via DefinePlugin in webpack configuration.
+  const NX_APP = /^NX_/i;
+
+  const raw = Object.keys(process.env)
+    .filter((key) => NX_APP.test(key))
+    .reduce(
+      (env, key) => {
+        env[key] = process.env[key];
+        return env;
+      },
+      // If mode is undefined or is dev or prod then webpack already defines this variable for us.
+      !mode || mode === 'development' || mode === 'production'
+        ? {}
+        : {
+            NODE_ENV: process.env.NODE_ENV,
+          }
+    );
+
+  // Stringify all values so we can feed into webpack DefinePlugin
+  const stringified = {
+    'process.env': Object.keys(raw).reduce((env, key) => {
+      env[key] = JSON.stringify(raw[key]);
+      return env;
+    }, {}),
+  };
+
+  return { stringified };
+}
