+++ packages/webpack/src/utils/webpack/package-chunk-sort.ts
@@ -13,9 +13,7 @@ export function generateEntryPoints(appConfig: {
     const entryPoints = normalizeExtraEntryPoints(
       extraEntryPoints,
       defaultBundleName
-    )
-      .filter((entry) => entry.inject)
-      .map((entry) => entry.bundleName);
+    ).map((entry) => entry.bundleName);
 
     // remove duplicates
     return [...new Set(entryPoints)];
