--- packages/webpack/src/utils/webpack/plugins/raw-css-loader.ts
@@ -1,5 +1,0 @@
-export default function RawCssLoader(content: string, map: object) {
-  const stringifiedContent = JSON.stringify(content);
-  const stringifiedMap = map ? JSON.stringify(map) : `''`;
-  return `module.exports = [[module.id, ${stringifiedContent}, '', ${stringifiedMap}]]`;
-}
